Ea_OS/Signals/quenyan/pyproject.toml
[project]
name = "qyn1"
version = "0.1.0"
description = "Reference implementation of the Quenya Morphemic Crypto-Language (QYN-1) pipeline"
authors = [
    { name = "Autogenerated by OpenAI ChatGPT" }
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10,<4.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pyyaml>=6.0,<7.0",
    "cryptography>=42.0.0,<43.0.0",
    "argon2-cffi>=23.1.0,<24.0.0",
]

[project.urls]
Homepage = "https://example.com/quenyan"
Documentation = "https://example.com/quenyan/docs"
Repository = "https://example.com/quenyan/repository"

[project.optional-dependencies]
dev = [
    "build>=1.2",
    "coverage[toml]>=7.4",
    "cyclonedx-bom>=3.7",
    "flake8>=7.0",
    "hypothesis>=6.90",
    "mypy>=1.9",
    "nox>=2024.4.15",
    "pip-audit>=2.7",
    "pip-tools>=7.4",
    "pytest>=7.4",
    "ruff>=0.4.7",
    "sigstore>=3.0",
    "tox>=4.15",
    "twine>=5.0",
    "types-PyYAML>=6.0.12.12",
]
audit = ["pip-audit>=2.7"]
sbom = ["cyclonedx-bom>=3.7"]
release = [
    "build>=1.2",
    "cyclonedx-bom>=3.7",
    "pip-audit>=2.7",
    "sigstore>=3.0",
    "twine>=5.0",
]
ci = ["nox>=2024.4.15", "tox>=4.15"]

[project.scripts]
quenyan = "qyn1.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
branch = true
source = ["qyn1"]

[tool.coverage.report]
fail_under = 80
skip_covered = true
show_missing = true

[tool.hypothesis]
deadline = 1000
max_examples = 50

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
follow_imports = "silent"
warn_unused_ignores = true
namespace_packages = true
allow_untyped_defs = true
allow_subclassing_any = true
exclude = ["tests/data"]
warn_unused_configs = true
warn_redundant_casts = true

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["qyn1", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203", "E501"]
extend-select = ["RUF100"]

[tool.pytest.ini_options]
addopts = "-q --strict-markers"
markers = [
    "property: property-based fuzzing suites",
]
python_files = "test_*.py"
